pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gp-chat"
version = "0.1.4" 
description = "GP-chat on GCP"
authors = [
    { name = "yoichi-1984", email = "yoichi.1984.engineer@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "streamlit==1.52.2",
    "google-genai==1.56.0",
    "python-dotenv==1.2.1",
    "streamlit-ace==0.1.1",
    "build==1.3.0",
    "pylint==4.0.4",
    "PyYAML==6.0.3",
    "python-docx==1.2.0",
    "python-pptx==1.0.2",
    "fastapi==0.129.2",
    "uvicorn==0.41.0",
    "firebase-admin==7.1.0",
    "httpx==0.28.1",
    "websockets==15.0.1",
    "cryptography",
    "google-cloud-firestore",    
    "streamlit-paste-button",
    "pandas",
    "numpy",
    "matplotlib",
    "google-cloud-storage"
]

[project.scripts]
gp-chat = "gp_chat.main_runner:run"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
gp_chat = ["*.yaml"]